backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
python-multipart==0.0.20
python-docx==1.1.2
